# core: fix and speed up mempool Verify()